+++ init/archlinux/README.md
@@ -1,0 +1,1 @@
+This template file is built by GitHub Actions and uploaded to a [different github repo](https://github.com/golift/aur).
