+++ init/bsd/freebsd.rc.d
@@ -1,31 +1,31 @@
 #!/bin/sh
 #
-# FreeBSD rc.d startup script for {{BINARY}}.
+# FreeBSD rc.d startup script for turbovanityurls.
 #
-# PROVIDE: {{BINARY}}
+# PROVIDE: turbovanityurls
 # REQUIRE: networking syslog
 # KEYWORD:
 
 . /etc/rc.subr
 
-name="{{BINARYU}}"
-real_name="{{BINARY}}"
-rcvar="{{BINARYU}}_enable"
-{{BINARYU}}_command="/usr/local/bin/${real_name}"
-{{BINARYU}}_user="{{BINARY}}"
-{{BINARYU}}_config="/usr/local/etc/${real_name}/{{CONFIG_FILE}}"
+name="turbovanityurls"
+real_name="turbovanityurls"
+rcvar="turbovanityurls_enable"
+turbovanityurls_command="/usr/local/bin/${real_name}"
+turbovanityurls_user="turbovanityurls"
+turbovanityurls_config="/usr/local/etc/${real_name}/config.yaml"
 pidfile="/var/run/${real_name}/pid"
 
-# This runs `daemon` as the `{{BINARYU}}_user` user.
+# This runs `daemon` as the `turbovanityurls_user` user.
 command="/usr/sbin/daemon"
-command_args="-P ${pidfile} -r -t ${real_name} -T ${real_name} -l daemon ${{{BINARYU}}_command} -c ${{{BINARYU}}_config}"
+command_args="-P ${pidfile} -r -t ${real_name} -T ${real_name} -l daemon ${turbovanityurls_command} -c ${turbovanityurls_config}"
 
 load_rc_config ${name}
-: ${{{BINARYU}}_enable:=no}
+: ${turbovanityurls_enable:=no}
 
 # Make a place for the pid file.
 mkdir -p $(dirname ${pidfile})
-chown -R ${{BINARYU}}_user $(dirname ${pidfile})
+chown -R $turbovanityurls_user $(dirname ${pidfile})
 
 # Suck in optional exported override variables.
 # ie. add something like the following to this file: export UP_POLLER_DEBUG=true
