--- init/docker/build
@@ -1,42 +1,0 @@
-#!/bin/bash
-
-# This always run local to the Dockerfile folder, so the path is ../..
-pushd ../..
-
-source settings.sh
-
-docker buildx create --use
-
-if [ "v$VERSION" = "$SOURCE_BRANCH" ]; then
-  TAGS="--tag ${DOCKER_REPO}:$VERSION"
-
-  echo $SOURCE_BRANCH | grep -q -- -
-  if [ "$?" = "1" ]; then
-    # tag does not contain a dash, so assume it's a prod tag.
-    TAGS="$TAGS --tag ${DOCKER_REPO}:latest --tag ${DOCKER_REPO}:stable --tag ${DOCKER_REPO}:$(echo $VERSION | cut -d. -f1,2) --tag ${DOCKER_REPO}:$(echo $VERSION | cut -d. -f1)"
-  fi
-else
-  TAGS="--tag ${DOCKER_REPO}:${SOURCE_BRANCH}"
-fi
-
-
-read -r -d '' CMD << EOF 
- docker buildx build --push ${TAGS} \
-  --platform linux/arm/v7,linux/arm64/v8,linux/amd64 \
-  --build-arg "BUILD_DATE=${DATE}" \
-  --build-arg "BUILD_FLAGS=${BUILD_FLAGS}" \
-  --build-arg "COMMIT=${COMMIT}" \
-  --build-arg "VERSION=${VERSION}-${ITERATION}" \
-  --build-arg "LICENSE=${LICENSE}" \
-  --build-arg "DESC=${DESC}" \
-  --build-arg "VENDOR=${VENDOR}" \
-  --build-arg "AUTHOR=${MAINT}" \
-  --build-arg "BINARY=${BINARY}" \
-  --build-arg "SOURCE_URL=${SOURCE_URL}" \
-  --file ${DOCKERFILE_PATH} .
-EOF
-
-echo "Running command: ${CMD}"
-eval $CMD
-
-popd
