+++ init/systemd/before-remove.sh
@@ -1,0 +1,12 @@
+#!/bin/bash
+
+# This file is used by rpm and deb packages. FPM use.
+
+if [ "$1" = "upgrade" ] || [ "$1" = "1" ] ; then
+  exit 0
+fi
+
+if [ -x "/bin/systemctl" ]; then
+  /bin/systemctl stop turbovanityurls
+  /bin/systemctl disable turbovanityurls
+fi
