+++ init/systemd/turbovanityurls.service
@@ -1,28 +1,27 @@
-# Systemd service unit for {{BINARY}}.
+# Systemd service unit for turbovanityurls.
 # This is part of Application Builder.
 # https://github.com/golift/application-builder
-# This file is installed when FORMULA is set to 'service'.
 
 [Unit]
-Description={{BINARY}} - {{DESC}}
+Description=turbovanityurls - Go Vanity URL Server
 After=network.target
 Requires=network.target
 
 [Service]
-ExecStart=/usr/bin/{{BINARY}} $DAEMON_OPTS
-EnvironmentFile=-/etc/default/{{BINARY}}
-EnvironmentFile=-/etc/sysconfig/{{BINARY}}
+ExecStart=/usr/bin/turbovanityurls $DAEMON_OPTS
+EnvironmentFile=-/etc/default/turbovanityurls
+EnvironmentFile=-/etc/sysconfig/turbovanityurls
 Restart=always
 RestartSec=10
 StandardOutput=syslog
 StandardError=syslog
-SyslogIdentifier={{BINARY}}
+SyslogIdentifier=turbovanityurls
 Type=simple
 
 # These should be set correctly for your environment.
 UMask=0002
-User={{BINARY}}
-Group={{BINARY}}
+User=turbovanityurls
+Group=turbovanityurls
 
 [Install]
 WantedBy=multi-user.target
