+++ pkg/service/build_darwin.go
@@ -1,7 +1,7 @@
-//go:build freebsd
-// +build freebsd
+//go:build darwin
+// +build darwin
 
-package main
+package service
 
 // DefaultConfFile is where to find config if -c is not provided.
 const DefaultConfFile = `/usr/local/etc/turbovanityurls/config.yaml`
