+++ pkg/service/build_freebsd.go
@@ -1,7 +1,7 @@
-//go:build darwin
-// +build darwin
+//go:build freebsd
+// +build freebsd
 
-package main
+package service
 
 // DefaultConfFile is where to find config if -c is not provided.
 const DefaultConfFile = `/usr/local/etc/turbovanityurls/config.yaml`
