+++ pkg/service/build_windows.go
@@ -1,7 +1,7 @@
 //go:build windows
 // +build windows
 
-package main
+package service
 
 // DefaultConfFile is where to find config if -c is not prvided.
 const DefaultConfFile = `C:\ProgramData\turbovanityurls\config.yaml`
