+++ settings.sh
@@ -3,34 +3,31 @@
 # Powered by Application Builder: https://github.com/golift/application-builder
 
 # Bring in dynamic repo/pull/source info.
-source $(dirname "${BASH_SOURCE[0]}")/init/buildinfo.sh
-
-# Must match the repo name to make things easy. Otherwise, fix some other paths.
-BINARY="turbovanityurls"
-REPO="golift/turbovanityurls"
-# Github repo containing homebrew formula repo.
-HBREPO="golift/homebrew-mugs"
 MAINT="David Newhall II <captain at golift dot io>"
 DESC="HTTP Server providing vanity go import paths."
-# Example must exist at examples/$CONFIG_FILE.example
-CONFIG_FILE="config.yaml"
 LICENSE="Apache-2.0"
-# FORMULA is either 'service' or 'tool'. Services run as a daemon, tools do not.
-# This affects the homebrew formula (launchd) and linux packages (systemd).
-FORMULA="service"
+SOURCE_URL="https://github.com/golift/turbovanityurls"
+VENDOR="Go Lift <code@golift.io>"
+export MAINT DESC LICENSE SOURCE_URL VENDOR
+
+DATE="$(date -u +%Y-%m-%dT%H:%M:00Z)"
+VERSION=$(git describe --abbrev=0 --tags $(git rev-list --tags --max-count=1) | tr -d v)
+[ "$VERSION" != "" ] || VERSION=development
+# This produces a 0 in some environments (like Homebrew), but it's only used for packages.
+ITERATION=$(git rev-list --count --all || echo 0)
+COMMIT="$(git rev-parse --short HEAD || echo 0)"
+GIT_BRANCH="$(git rev-parse --abbrev-ref HEAD || echo unknown)"
+BRANCH="${GIT_BRANCH:-${GITHUB_REF_NAME}}"
+export DATE VERSION ITERATION COMMIT BRANCH
+
+
 
 # Used for source links and wiki links.
-SOURCE_URL="https://github.com/${REPO}"
-
-# This parameter is passed in as -X to go build. Used to override the Version variable in a package.
-# This makes a path like github.com/user/hello-world/helloworld.Version=1.3.3
-# Name the Version-containing library the same as the github repo, without dashes.
-VERSION_PATH="main"
 
 # Used by homebrew downloads.
 SOURCE_PATH=https://codeload.github.com/${REPO}/tar.gz/v${VERSION}
 
-export BINARY GHUSER HBREPO MAINT DESC CONFIG_FILE
+export GHUSER MAINT DESC CONFIG_FILE
 export LICENSE FORMULA SOURCE_URL VERSION_PATH SOURCE_PATH
 
 ### Optional ###
@@ -47,6 +44,6 @@ export EXTRA_FPM_FLAGS=""
 # These do not affect automated builds, just allow the docker build scripts to run from a local clone.
 [ -n "$SOURCE_BRANCH" ] || export SOURCE_BRANCH=$BRANCH
 [ -n "$DOCKER_TAG" ] || export DOCKER_TAG=$(echo $SOURCE_BRANCH | sed 's/^v*\([0-9].*\)/\1/')
-[ -n "$DOCKER_REPO" ] || export DOCKER_REPO="golift/${BINARY}"
+[ -n "$DOCKER_REPO" ] || export DOCKER_REPO="golift/turbovanityurls"
 [ -n "$IMAGE_NAME" ] || export IMAGE_NAME="${DOCKER_REPO}:${DOCKER_TAG}"
 [ -n "$DOCKERFILE_PATH" ] || export DOCKERFILE_PATH="init/docker/Dockerfile"
